+++ rules-tests/CodeQuality/Rector/MethodCall/MatchAssertEqualsExpectedTypeRector/Fixture/match_integer_to_string.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\PHPUnit\Tests\CodeQuality\Rector\MethodCall\MatchAssertSameExpectedTypeRector\Fixture;
+namespace Rector\PHPUnit\Tests\CodeQuality\Rector\MethodCall\MatchAssertEqualsExpectedTypeRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
 
@@ -8,7 +8,7 @@ final class MatchIntegerToString extends TestCase
 {
     public function test()
     {
-        $this->assertSame(123, $this->getOrderId());
+        $this->assertEquals(123, $this->getOrderId());
     }
 
     private function getOrderId(): string
@@ -21,7 +21,7 @@ final class MatchIntegerToString extends TestCase
 -----
 <?php
 
-namespace Rector\PHPUnit\Tests\CodeQuality\Rector\MethodCall\MatchAssertSameExpectedTypeRector\Fixture;
+namespace Rector\PHPUnit\Tests\CodeQuality\Rector\MethodCall\MatchAssertEqualsExpectedTypeRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
 
@@ -29,7 +29,7 @@ final class MatchIntegerToString extends TestCase
 {
     public function test()
     {
-        $this->assertSame('123', $this->getOrderId());
+        $this->assertEquals('123', $this->getOrderId());
     }
 
     private function getOrderId(): string
