+++ rules-tests/CodeQuality/Rector/MethodCall/MatchAssertEqualsExpectedTypeRector/Fixture/skip_first_class_callable.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\PHPUnit\Tests\CodeQuality\Rector\MethodCall\MatchAssertSameExpectedTypeRector\Fixture;
+namespace Rector\PHPUnit\Tests\CodeQuality\Rector\MethodCall\MatchAssertEqualsExpectedTypeRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
