--- rules-tests/CodeQuality/Rector/MethodCall/MatchAssertSameExpectedTypeRector/config/configured_rule.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-use Rector\PHPUnit\CodeQuality\Rector\MethodCall\MatchAssertSameExpectedTypeRector;
-
-return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->rule(MatchAssertSameExpectedTypeRector::class);
-};
