+++ rules/CodeQuality/Rector/MethodCall/MatchAssertEqualsExpectedTypeRector.php
@@ -16,9 +16,9 @@ use Symplify\RuleDocGenerator\ValueObject\CodeSample\CodeSample;
 use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
 
 /**
- * @see \Rector\PHPUnit\Tests\CodeQuality\Rector\MethodCall\MatchAssertSameExpectedTypeRector\MatchAssertSameExpectedTypeRectorTest
+ * @see \Rector\PHPUnit\Tests\CodeQuality\Rector\MethodCall\MatchAssertEqualsExpectedTypeRector\MatchAssertEqualsExpectedTypeRectorTest
  */
-final class MatchAssertSameExpectedTypeRector extends AbstractRector
+final class MatchAssertEqualsExpectedTypeRector extends AbstractRector
 {
     public function __construct(
         private readonly TestsNodeAnalyzer $testsNodeAnalyzer
@@ -28,7 +28,7 @@ final class MatchAssertSameExpectedTypeRector extends AbstractRector
     public function getRuleDefinition(): RuleDefinition
     {
         return new RuleDefinition(
-            'Correct expected type in assertSame() method to match strict type of passed variable',
+            'Correct expected type in assertEquals() method to match strict type of passed variable',
             [
                 new CodeSample(
                     <<<'CODE_SAMPLE'
@@ -38,7 +38,7 @@ class SomeTest extends TestCase
 {
     public function run()
     {
-        $this->assertSame('123', $this->getOrderId());
+        $this->assertEquals('123', $this->getOrderId());
     }
 
     private function getOrderId(): int
@@ -55,7 +55,7 @@ class SomeTest extends TestCase
 {
     public function run()
     {
-        $this->assertSame(123, $this->getOrderId());
+        $this->assertEquals(123, $this->getOrderId());
     }
 
     private function getOrderId(): int
@@ -82,7 +82,7 @@ CODE_SAMPLE
      */
     public function refactor(Node $node): ?Node
     {
-        if (! $this->testsNodeAnalyzer->isPHPUnitMethodCallNames($node, ['assertSame', 'assertEquals'])) {
+        if (! $this->testsNodeAnalyzer->isPHPUnitMethodCallNames($node, ['assertEquals'])) {
             return null;
         }
